Begin restructure
Moved all js files into src/services/discord, created directories and files to organize the behemoth of index.js
