--- banlist.txt
@@ -1,3 +1,0 @@
-154643218135246456
-3215649854132165
-125465461531321654
