+++ botconfig.json
@@ -1,4 +1,4 @@
 {
-"token": "secret",
+"token": "MTAyMjY0MTU0ODU3MzI5MDUxNw.G8DBlw.fhpNiqs9hXVcXnpjkqWORt0WvRd4rmJzeN5wDI",
 "prefix": "."
 }
